backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbilens-backend"
version = "0.1.0"
description = "FastAPI backend for ArbiLens"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "boto3>=1.35.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "openai>=1.54.0",
  "PyJWT>=2.10.0",
  "passlib[bcrypt]>=1.7.4",
  "pypdf>=5.4.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "python-docx>=1.1.2",
  "python-json-logger>=2.0.7",
  "python-multipart>=0.0.12",
  "reportlab>=4.2.5",
  "stripe>=11.4.0",
  "tenacity>=9.0.0",
  "redis>=5.2.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.8.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
